Refactoring and splitting the code
